TaylorInterface.jl

Documentation under construction.

Main structures :

TaylorInterface.TaylorGenerator — Type

TaylorGenerator(name=randstring(12), eqs_filename, parent_path=".")

Structure used to generate directories and call taylor. The directory name will be $(parent_path)/taylor_$(name). The equation file should follow the syntax required by taylor.

TaylorInterface.TaylorHandler — Type

TaylorHandler(path)

Structure used to open, use and close directories generated by TaylorGenerator.