Full documentation

Main structures

TaylorInterface.TaylorGenerator — Type

TaylorGenerator(name=randstring(12), eqs_filename, parent_path=".")

Structure used to generate directories and call taylor. The directory name will be $(parent_path)/taylor_$(name). The equation file should follow the syntax required by taylor.

TaylorInterface.TaylorHandler — Type

TaylorHandler(path)

Structure used to open, use and close directories generated by TaylorGenerator.

Around the TaylorGenerator

TaylorInterface.generate_dir — Function

generate_dir(generator)

Generates the directory corresponding to the generator, and builds the library.

TaylorInterface.get_taylor_dir — Function

get_taylor_dir(generator)

Returns the directory associated to the generator.

TaylorInterface.get_handler — Function

get_handler(generator, open=true)

Get the TaylorHandler corresponding to the generator. If open is true, the library will also be open in the handler.

Around the TaylorHandler

TaylorInterface.compile_eqs — Function

compile_eqs(generator)

Runs make in the directory of the generator.

compile_eqs(handler)

Runs make in the directory of the handler.

TaylorInterface.open_lib — Function

open_lib(handler)

Opens the handler's library, storing the loaded library and a dictionary with symbols to functions in the handler.

Currently, the loaded functions are flow, tstep and tstep_reverse.

TaylorInterface.close_lib — Function

close_lib(handler)

Close the library referred by the handler.

TaylorInterface.clear_dir — Function

clear_dir(handler)

Delete the directory referred by the handler.

clear_dir(generator)

Delete the directory referred by the generator.

TaylorInterface.is_open — Function

is_open(handler)

Check if handler refers to an open library.

TaylorInterface.check_open — Function

check_open(handler)

Check if handler refers to an open library, and opens it if not.

Default models

TaylorInterface.get_model — Function

get_model(name, open=false)

Get the TaylorHandler corresponding to a default model in this package. If open is true, the library will also be open in the handler.

Warning: name should NOT contain the prefix "taylor_h".

Internal functions

File system navigation

These functions are mainly intended for internal use and navigation within the file system.

TaylorInterface.get_dir_src — Function

get_dir_src()

Returns the directory containing this package.

TaylorInterface.get_models_dir — Function

get_models_dir()

Returns the directory containing predefined models.

TaylorInterface.get_wrappers — Function

get_wrappers(generator)

Get the appropriate wrapper files for the generator.

Taylor flags

TaylorInterface.get_jet_flag — Function

get_jet_flag(generator)

Get the jet flag to pass to taylor when compiling the equations given by the generator.